sync change to sonic_platform_base/sonic_sfp and create symbol link
