+++ sonic_platform_base/sonic_sfp/sff8024.py
@@ -34,3 +34,33 @@ type_of_transceiver = {
     '19': 'OSFP 8X Pluggable Transceiver',
     '1a': 'SFP-DD Double Density 2X Pluggable Transceiver'
 }
+
+type_abbrv_name = {
+    '00': 'Unknown',
+    '01': 'GBIC',
+    '02': 'Soldered',
+    '03': 'SFP',
+    '04': 'XBI300',
+    '05': 'XENPAK',
+    '06': 'XFP',
+    '07': 'XFF',
+    '08': 'XFP-E',
+    '09': 'XPAK',
+    '0a': 'X2',
+    '0b': 'DWDM-SFP',
+    '0c': 'QSFP',
+    '0d': 'QSFP+',
+    '0e': 'CXP',
+    '0f': 'HD4X',
+    '10': 'HD8X',
+    '11': 'QSFP28',
+    '12': 'CXP2',
+    '13': 'CDFP-1/2',
+    '14': 'HD4X-Fanout',
+    '15': 'HD8X-Fanout',
+    '16': 'CDFP-3',
+    '17': 'MicroQSFP',
+    '18': 'QSFP-DD',
+    '19': 'OSFP-8X',
+    '1a': 'SFP-DD'
+}
