+++ sonic_sfp
@@ -1,0 +1,1 @@
+./sonic_platform_base/sonic_sfp/
